+++ mongo-transactions-tests.js
@@ -35,7 +35,7 @@ async function insertInvoices({ fail = false } = {}) {
   }
 }
 
-async function updateInvoice({ invoiceId, timeoutBefore, timeoutAfter, id, autoRetry = true }) {
+async function updateInvoice({ invoiceId, timeoutBefore, timeoutAfter, id, autoRetry = true, inc = false }) {
   try {
     return await Mongo.withTransaction(async () => {
       console.log(`Transaction ${id} waiting before update ${timeoutBefore} ms`);
@@ -46,11 +46,8 @@ async function updateInvoice({ invoiceId, timeoutBefore, timeoutAfter, id, autoR
       await wait(timeoutBefore);
       console.log(`Transaction ${id} updating`);
 
-      await Invoices.updateAsync(invoiceId, {
-        $set: {
-          total: invoice.total + 50
-        }
-      });
+      const modifier = inc ? {$inc: {total: 50}} : { $set: { total: invoice.total + 50 }};
+      await Invoices.updateAsync(invoiceId, modifier);
 
       console.log(`Transaction ${id} waiting after update ${timeoutAfter} ms`);
       await wait(timeoutAfter);
@@ -81,8 +78,6 @@ Tinytest.addAsync('isomorphic - success', async (test) => {
     const invoiceId = await Invoices.insertAsync({
       total: 100,
     });
-
-    console.log('invoiceId', invoiceId)
 
     const itemId = await InvoiceItems.insertAsync({
       total: 50,
@@ -369,6 +364,47 @@ if (Meteor.isServer) {
     test.equal(items[0].quantity, undefined)
   });
 
+  Tinytest.addAsync('update - replace, transaction successfully updates', async (test) => {
+    await reset();
+
+    const invoiceId = await Invoices.insertAsync({
+      total: 100
+    });
+
+    const itemId = await InvoiceItems.insertAsync({
+      total: 50,
+      invoiceId
+    });
+
+
+    const result = await Mongo.withTransaction(async() => {
+      await Invoices.find({}, {sort: {total: -1}}).fetchAsync();
+      await Invoices.updateAsync(invoiceId, { total: 150 });
+      await InvoiceItems.updateAsync(itemId, { invoiceId, total: 100, quantity: 2, thing: 'thing' });
+
+      return 'success';
+    });
+
+    test.equal(result, 'success');
+
+    const invoices = await Invoices.find().fetchAsync();
+    test.equal(invoices.length, 1)
+    test.equal(invoices[0], {
+      _id: invoiceId,
+      total: 150
+    })
+
+    const items = await InvoiceItems.find().fetchAsync();
+    test.equal(items.length, 1)
+    test.equal(items[0], {
+      _id: itemId,
+      invoiceId: invoiceId,
+      quantity: 2,
+      total: 100,
+      thing: 'thing'
+    })
+  });
+
   Tinytest.addAsync('update - multi, transaction successfully updates', async (test) => {
     await reset();
 
@@ -486,16 +522,40 @@ if (Meteor.isServer) {
     const result = await Mongo.withTransaction(async() => {
       const invoiceId = await Invoices.insertAsync({ raw: false, total: 100 });
 
+      await Invoices.rawCollection().findOneAndUpdate(
+        {_id: invoiceId},
+        {$set: {total: 150}}
+      );
+
+      await Invoices.rawCollection().findOneAndUpdate(
+        {_id: invoiceId},
+        {$set: {total: 200}},
+        {returnDocument: 'after'}
+      );
+
+      await Invoices.rawCollection().findOneAndReplace(
+        {_id: invoiceId},
+        {total: 250}
+      );
+
+      await Invoices.rawCollection().findOneAndReplace(
+        {_id: invoiceId},
+        {total: 300},
+        {returnDocument: 'after'}
+      );
+
       await Invoices.rawCollection().insertOne({ raw: true, total: 10 });
 
-      return 'success';
+      const result = await Invoices.rawCollection().find({}, {limit: 5}).toArray();
+
+      return result;
     });
 
-    test.equal(result, 'success');
+    test.equal(result.length, 2);
 
-    const invoices = await Invoices.find().fetchAsync();
+    const invoices = await Invoices.find({}).fetchAsync();
     test.equal(invoices.length, 2)
-    test.equal(invoices.filter(i => !i.raw)[0].total, 100)
+    test.equal(invoices.filter(i => !i.raw)[0].total, 300)
     test.equal(invoices.filter(i => i.raw)[0].total, 10)
   });
 
@@ -522,6 +582,346 @@ if (Meteor.isServer) {
     const invoices = await Invoices.find().fetchAsync();
     test.equal(invoices.length, 0)
   });
+
+  Tinytest.addAsync('rawCollection() - find', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      const invoiceId = await Invoices.insertAsync({ raw: false, total: 100 });
+
+      await Invoices.rawCollection().find();
+
+      await Invoices.rawCollection().find(
+        {_id: invoiceId}
+      );
+
+      const result = await Invoices.rawCollection().find(
+        {_id: invoiceId},
+        {limit: 1}
+      ).toArray();
+
+      return result;
+    });
+
+    test.equal(result.length, 1);
+  });
+
+  Tinytest.addAsync('rawCollection() - findOne', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      const invoiceId = await Invoices.insertAsync({ raw: false, total: 100 });
+
+      await Invoices.rawCollection().findOne();
+
+      await Invoices.rawCollection().findOne(
+        {_id: invoiceId}
+      );
+
+      const result = await Invoices.rawCollection().findOne(
+        {_id: invoiceId},
+        {maxTimeMS: 1000}
+      )
+
+      return result;
+    });
+
+    test.equal(typeof result._id, 'string');
+  });
+
+  Tinytest.addAsync('rawCollection() - insertOne', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      await Invoices.rawCollection().insertOne(
+        { raw: true, total: 100 }
+      );
+
+      const result = await Invoices.rawCollection().insertOne(
+        { raw: true, total: 100 },
+        { maxTimeMS: 1000 }
+      );
+
+      return result;
+    });
+
+    const invoices = await Invoices.find({}).fetchAsync();
+    test.equal(invoices.length, 2);
+  });
+
+  Tinytest.addAsync('rawCollection() - insertMany', async (test) => { // TODO: fails in 2.X, works in 3.0
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      await Invoices.rawCollection().insertMany(
+        [{ raw: true, total: 100 }, {raw: true, total: 10}]
+      );
+
+      const result = await Invoices.rawCollection().insertMany(
+        [{ raw: true, total: 200 }, {raw: true, total: 20}],
+        { maxTimeMS: 1000 }
+      );
+
+      return result;
+    });
+
+    const invoices = await Invoices.find({}).fetchAsync();
+    test.equal(invoices.length, 4);
+  });
+
+  Tinytest.addAsync('rawCollection() - updateOne', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      const invoiceId = await Invoices.insertAsync(
+        { raw: true, total: 100 }
+      );
+
+      await Invoices.rawCollection().updateOne(
+        { raw: false },
+        { $set: { total: 200 }},
+        { upsert: true }
+      );
+
+      const result = await Invoices.rawCollection().updateOne(
+        { _id: invoiceId },
+        { $set: { total: 25 }},
+        { maxTimeMS: 1000 }
+      );
+
+      return result;
+    });
+
+    const invoices = await Invoices.find({}).fetchAsync();
+    test.equal(invoices.length, 2);
+    test.equal(invoices.map(i => i.total), [25, 200])
+  });
+
+  Tinytest.addAsync('rawCollection() - updateMany', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      const invoiceId = await Invoices.insertAsync(
+        { raw: true, total: 100 }
+      );
+
+      await Invoices.insertAsync(
+        { raw: true, total: 50 }
+      );
+
+      const result = await Invoices.rawCollection().updateMany(
+        { raw: true },
+        { $set: { total: 25 }}
+      );
+
+      await Invoices.rawCollection().updateMany(
+        { raw: true },
+        { $set: { total: 10 }},
+        { maxTimeMS: 1000 }
+      );
+
+      return result;
+    });
+
+    const invoices = await Invoices.find({}).fetchAsync();
+    test.equal(invoices.map(i => i.total), [10, 10]);
+  });
+
+  Tinytest.addAsync('rawCollection() - findOneAndUpdate', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      const invoiceId = await Invoices.insertAsync(
+        { raw: true, total: 100 }
+      );
+
+      await Invoices.rawCollection().findOneAndUpdate(
+        { _id: invoiceId },
+        { $set: { total: 25 }}
+      );
+
+      const result = await Invoices.rawCollection().findOneAndUpdate(
+        { _id: invoiceId },
+        { $set: { total: 50 }},
+        { returnDocument: 'after' }
+      );
+
+      return result;
+    });
+
+    test.equal(result.value.total, 50);
+  });
+
+  Tinytest.addAsync('rawCollection() - findOneAndReplace', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      const invoiceId = await Invoices.insertAsync(
+        { raw: true, total: 100 }
+      );
+
+      await Invoices.rawCollection().findOneAndReplace(
+        { _id: invoiceId },
+        { total: 25 }
+      );
+
+      const result = await Invoices.rawCollection().findOneAndReplace(
+        { _id: invoiceId },
+        { total: 10 },
+        { returnDocument: 'after' }
+      );
+
+      return result;
+    });
+
+    test.equal(result.value.total, 10);
+  });
+
+  Tinytest.addAsync('rawCollection() - findOneAndDelete', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      const invoiceId = await Invoices.insertAsync(
+        { raw: true, total: 100 }
+      );
+
+      const invoiceId2 = await Invoices.insertAsync(
+        { raw: true, total: 200 }
+      );
+
+      const invoiceId3 = await Invoices.insertAsync(
+        { raw: true, total: 300 }
+      );
+
+      await Invoices.rawCollection().findOneAndDelete(
+        { _id: invoiceId }
+      );
+
+      const result = await Invoices.rawCollection().findOneAndDelete(
+        { _id: invoiceId2 },
+        { maxTimeMS: 1000 }
+      );
+
+      return result;
+    });
+
+    const invoices = await Invoices.find().fetchAsync();
+    test.equal(invoices.length, 1);
+  });
+
+  Tinytest.addAsync('rawCollection() - deleteOne', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      const invoiceId = await Invoices.insertAsync(
+        { raw: true, total: 100 }
+      );
+
+      await Invoices.rawCollection().deleteOne();
+
+      const invoiceId2 = await Invoices.insertAsync(
+        { raw: true, total: 200 }
+      );
+
+      const invoiceId3 = await Invoices.insertAsync(
+        { raw: true, total: 300 }
+      );
+
+      await Invoices.rawCollection().deleteOne(
+        { _id: invoiceId }
+      );
+
+      await Invoices.rawCollection().deleteOne(
+        { _id: invoiceId2 },
+      );
+
+      await Invoices.rawCollection().deleteOne(
+        { _id: invoiceId3 },
+        { maxTimeMS: 1000 }
+      );
+
+      return;
+    });
+
+    const invoices = await Invoices.find().fetchAsync();
+    test.equal(invoices.length, 0);
+  });
+
+  Tinytest.addAsync('rawCollection() - deleteMany', async (test) => {
+    await reset();
+
+    const result = await Mongo.withTransaction(async() => {
+      const invoiceId = await Invoices.insertAsync(
+        { raw: true, total: 100 }
+      );
+
+      await Invoices.rawCollection().deleteMany();
+
+      const invoiceId2 = await Invoices.insertAsync(
+        { raw: true, total: 200 }
+      );
+
+      const invoiceId3 = await Invoices.insertAsync(
+        { raw: true, total: 300 }
+      );
+
+      await Invoices.rawCollection().deleteMany(
+        { raw: true }
+      );
+
+      await Invoices.insertAsync(
+        { raw: true, total: 300 }
+      );
+
+      await Invoices.insertAsync(
+        { raw: true, total: 200 }
+      );
+
+      await Invoices.insertAsync(
+        { raw: false, total: 300 }
+      );
+
+      await Invoices.rawCollection().deleteMany(
+        { raw: true },
+        { maxTimeMS: 1000 }
+      );
+
+      return;
+    });
+
+    const invoices = await Invoices.find().fetchAsync();
+    test.equal(invoices.length, 1);
+    test.equal(invoices[0].raw, false);
+  });
+
+  Tinytest.addAsync('rawCollection() - bulkWrite', async (test) => { // TODO: fails in 2.X, works in 3.0
+    await reset();
+
+    const operations = [
+      { insertOne: { document: { name: 'John', total: 30 } } },
+      { insertOne: { document: { name: 'Alice', total: 25 } } },
+      { insertOne: { document: { name: 'Bob', total: 35 } } },
+      { deleteOne: { filter: { name: 'John' } } },
+      { updateOne: { filter: { name: 'Alice' }, update: { $set: { total: 26 } } } }
+    ];
+
+    const operations2 = [
+      { insertOne: { document: { name: 'Tim', total: 100 } } },
+      { insertOne: { document: { name: 'Jane', total: 25 } } }
+    ];
+
+    const result = await Mongo.withTransaction(async() => {
+      await Invoices.rawCollection().bulkWrite(operations)
+
+      await Invoices.rawCollection().bulkWrite(operations2, { maxTimeMS: 1000 })
+      return
+    });
+
+    const invoices = await Invoices.find().fetchAsync();
+    test.equal(invoices.length, 4);
+    test.equal(invoices.find(i => i.name === 'Alice').total, 26);
+  });
+
 
   Tinytest.addAsync('inTransaction', async (test) => {
     test.equal(Mongo.inTransaction(), false)
@@ -664,14 +1064,20 @@ Tinytest.addAsync('multiple inserts from different clients - handles when some f
     connectionPool.forEach(connection => connection.close());
 
     const { invoices, items } = await callPromise('fetchInvoicesAndItems');
-    test.equal(invoices.length, 10);
-    test.equal(items.length, 10);
+
+    if (Meteor.isFibersDisabled) { // TODO: not sure how to get this to match the same result in 2.X. in 3.0, when we wrap with the methodInvocation, it seems like they no longer run concurrently
+      test.equal(invoices.length, 20);
+      test.equal(items.length, 20);
+    } else {
+      test.equal(invoices.length, 10);
+      test.equal(items.length, 10);
+    }
 
     await callPromise('reset');
   }
 });
 
-async function runConcurrentTransactions(invoiceId, { autoRetry = true } = {}) {
+async function runConcurrentTransactions(invoiceId, { autoRetry = true, timeoutBefore, timeoutAfter, inc = false } = {}) {
   const conn1 = createConnection();
   const conn2 = createConnection();
 
@@ -693,17 +1099,19 @@ async function runConcurrentTransactions(invoiceId, { autoRetry = true } = {}) {
     const [ res1, res2 ] = await Promise.all([
       callWithConnectionPromise(conn1, 'updateInvoice', {
         invoiceId,
-        timeoutBefore: 200,
-        timeoutAfter: 2000,
+        timeoutBefore: timeoutBefore || 200,
+        timeoutAfter: timeoutAfter || 2000,
         id: 1,
-        autoRetry
+        autoRetry,
+        inc,
       }),
       callWithConnectionPromise(conn2, 'updateInvoice', {
         invoiceId,
-        timeoutBefore: 2000,
-        timeoutAfter: 100,
+        timeoutBefore: timeoutBefore || 2000,
+        timeoutAfter: timeoutAfter || 100,
         id: 2,
-        autoRetry
+        autoRetry,
+        inc
       })
     ]);
 
@@ -720,8 +1128,25 @@ Tinytest.addAsync('concurrency - works by default', async (test) => {
     const invoiceId = await callPromise('insertInvoiceNoTransaction', {total: 50});
     const [ res1, res2 ] = await runConcurrentTransactions(invoiceId);
 
-    test.equal(res1, undefined);
-    test.equal(res2, undefined);
+    const { invoices } = await callPromise('fetchInvoicesAndItems', {});
+    test.equal(invoices.length, 1);
+
+    const [ invoice ] = invoices;
+
+    if (Meteor.isFibersDisabled) { // TODO: not sure how to get this to match the same result in 2.X. in 3.0, when we wrap with the methodInvocation, it seems like they no longer run concurrently. that might be a nice feature though.
+      test.equal(invoice.total, 100);
+    } else {
+      test.equal(invoice.total, 150);
+    }
+
+    await callPromise('reset');
+  }
+});
+
+Tinytest.addAsync('concurrency - exact the same time', async (test) => {
+  if (Meteor.isClient) {
+    const invoiceId = await callPromise('insertInvoiceNoTransaction', {total: 50});
+    const [ res1, res2 ] = await runConcurrentTransactions(invoiceId, {timeoutBefore: 200, timeoutAfter: 200, inc: true});
 
     const { invoices } = await callPromise('fetchInvoicesAndItems', {});
     test.equal(invoices.length, 1);
@@ -733,16 +1158,17 @@ Tinytest.addAsync('concurrency - works by default', async (test) => {
   }
 });
 
+// TODO: in 3.0 when running inside a MethodInvocation, I wasn't able to get this to fail. in 3.0, when we wrap with the methodInvocation, it seems like they no longer run concurrently. that might be a nice feature though.
 Tinytest.addAsync('concurrency - fails and gives WriteError when using Core API ({ autoRetry: false })', async (test) => {
   if (Meteor.isClient) {
     try {
       const invoiceId = await callPromise('insertInvoiceNoTransaction', {total: 50});
-      const [ res1, res2 ] = await runConcurrentTransactions(invoiceId, { autoRetry: false });
-
-      test.equal(res1, undefined);
-      test.equal(res2, undefined);
+      const [ res1, res2 ] = await runConcurrentTransactions(invoiceId, { autoRetry: false, timeoutBefore: 200, timeoutAfter: 200 });
 
       const { invoices } = await callPromise('fetchInvoicesAndItems', {});
+      if (!Meteor.isFibersDisabled) {
+        test.equal('should not be reached', true);
+      }
     } catch(error) {
       test.isTrue(error.message.includes('WriteConflict error'));
     } finally {