+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'jam:mongo-transactions',
-  version: '1.0.0',
+  version: '1.0.1',
   summary: 'An easy way to use Mongo Transactions for Meteor apps',
   git: 'https://github.com/jamauro/mongo-transactions',
   documentation: 'README.md'
@@ -10,6 +10,7 @@ Package.onUse(function(api) {
   api.versionsFrom(['2.8.1', '3.0-alpha.19']);
   api.use('ecmascript');
   api.use('mongo');
+  api.use('ddp-client');
   api.use('zodern:types@1.0.11');
   api.mainModule('mongo-transactions-client.js', 'client');
   api.mainModule('mongo-transactions-server.js', 'server');