+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 1.0.1
+* fix: preserve `MethodInvocation` when available
+* fix: better support for `rawCollection` methods
+
 ## 1.0
 * initial version